+++ packages/next-common/utils/consts/settings/kusama/index.js
@@ -41,7 +41,6 @@ const kusama = {
   showAchainableLabels: true,
   description:
     "Kusama is a scalable multi-chain network for radical innovation and early stage Polkadot deployments. Expect Chaos. No promises.",
-  hideNewTreasuryProposalButton: true,
   modules: {
     referenda: true,
     fellowship: true,