+++ packages/next-common/utils/consts/settings/polkadot/index.js
@@ -47,7 +47,6 @@ const polkadot = {
   showAchainableLabels: false,
   description:
     "Polkadot empowers blockchain networks to work together under the protection of shared security.",
-  hideNewTreasuryProposalButton: true,
   ...polkadotThemeVars,
   modules: {
     referenda: true,
@@ -58,6 +57,9 @@ const polkadot = {
     mimir: true,
   },
   treasuryProposalTracks: polkadotTreasuryTracks,
+  newProposalQuickStart: {
+    usdxTreasuryProposal: true,
+  },
 };
 
 export default polkadot;